removal: removed all ANN search methods that either did not perform well, or delivered bad results
This includes hora, hnsw_rs, and my custom ANN search using change hashing
