+++ Cargo.toml
@@ -14,14 +14,8 @@ log = "0.4.17"
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.9.14"
 derivative = "2.2.0"
-ngrammatic = "0.4.0"
-hora = "0.1.1"
 firestorm = { version= "0.5.1", features=["enable_system_time"] }
-hnsw_rs = "0.1.17"
 bit-vec = "0.6.3"
-num-traits = "0.2.15"
-threadpool = "1.8.1"
-faiss = "0.12.0"
 
 [dev-dependencies]
 criterion = "0.4.0"
