+++ src/search/methods.rs
@@ -1,4 +1,3 @@
-pub mod lsh;
 pub mod exact_diff;
+pub mod lsh;
 pub mod message_scan;
-pub mod similar_diff;
