--- src/search/util.rs
@@ -1,1 +1,0 @@
-pub mod ann;
